+++ repl/test/dotty/tools/repl/ReplDependencyMacroTests.scala
@@ -78,7 +78,7 @@ class ReplDependencyMacroTests extends ReplTest:
     initially:
       run(":dep com.lihaoyi::ujson:4.4.3")
       val depOutput = storedOutput()
-      assertTrue(depOutput, depOutput.contains("Resolved 1 dependencies"))
+      assertTrue(depOutput, depOutput.contains("Resolved a dependency"))
       assertNoMacroFailure(depOutput)
 
       run("ujson.Num(3.5)")
